+++ egui_web/src/lib.rs
@@ -15,6 +15,8 @@
 #![warn(clippy::all, missing_crate_level_docs, rust_2018_idioms)]
 
 pub mod backend;
+#[cfg(feature = "glow")]
+mod glow_wrapping;
 mod painter;
 pub mod screen_reader;
 pub mod webgl1;